examples/openclaw-monitoring/pyproject.toml
[project]
name = "openclaw-monitoring"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "pycloudflared>=0.3.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "videodb>=0.4.0",
]
